+++ examples/examples_nodejs_test.go
@@ -55,6 +55,15 @@ func TestWebserverTs(t *testing.T) {
 	integration.ProgramTest(t, &test)
 }
 
+func TestStorageTransferTs(t *testing.T) {
+	test := getJSBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir:         filepath.Join(getCwd(t), "storagetransfer-ts"),
+			ExpectRefreshChanges: true,
+		})
+	integration.ProgramTest(t, &test)
+}
+
 func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
 	base := getBaseOptions(t)
 	baseJS := base.With(integration.ProgramTestOptions{